rename attributes to attribute
