+++ koku/api/urls.py
@@ -44,14 +44,14 @@ from api.views import (
     billing_source,
     openapi
 )
-from attributes.api_views import AttributeViewSet
+from attribute.views import AttributeViewSet
 
 ROUTER = DefaultRouter()
 ROUTER.register(r'dataexportrequests', DataExportRequestViewSet, base_name='dataexportrequests')
 ROUTER.register(r'metrics', CostModelMetricsMapViewSet, base_name='metrics')
 ROUTER.register(r'providers', ProviderViewSet)
 ROUTER.register(r'preferences', UserPreferenceViewSet, base_name='preferences')
-ROUTER.register(r'attributes2', AttributeViewSet, base_name='attributes')
+ROUTER.register(r'attributes', AttributeViewSet, base_name='attributes')
 
 # pylint: disable=invalid-name
 urlpatterns = [
