+++ koku/attribute/urls.py
@@ -1,9 +1,9 @@
 from rest_framework import routers
-from .views import AttributeViewset
+from .views import AttributeViewSet
 from django.urls import include, path
 
 router = routers.DefaultRouter()
-router.register(r'attributes', AttributeViewset)
+router.register(r'attributes', AttributeViewSet)
 
 urlpatterns = [
     path('/', include(router.urls)),
